+++ internal/ui/main_view.go
@@ -52,11 +52,16 @@ func Make(a fyne.App, w fyne.Window) fyne.CanvasObject {
 	return mw.view
 }
 
+func (mw *mainView) setView(v fyne.CanvasObject) {
+	mw.view.Objects[0] = v
+	mw.view.Refresh()
+}
+
 func (mw *mainView) makeMainMenu() *fyne.MainMenu {
 	// a Quit item will is appended automatically by Fyne to the first menu item
 	fileMenu := fyne.NewMenu("File",
 		fyne.NewMenuItem("New Vault", func() {
-			mw.view.Objects[0] = mw.createVaultView()
+			mw.setView(mw.createVaultView())
 		}),
 	)
 	switchItem := fyne.NewMenuItem("Switch Vault", func() {
@@ -104,7 +109,7 @@ func (mw *mainView) makeMainMenu() *fyne.MainMenu {
 }
 
 func (mw *mainView) Reload() {
-	mw.view.Objects[0] = mw.buildMainView()
+	mw.setView(mw.buildMainView())
 	mw.SetMainMenu(mw.makeMainMenu())
 }
 
@@ -140,18 +145,13 @@ func (mw *mainView) initVaultView() fyne.CanvasObject {
 	password.SetPlaceHolder("Password")
 
 	btn := widget.NewButton("Create Vault", func() {
-		key, err := paw.NewKey(name.Text, password.Text)
-		if err != nil {
-			dialog.ShowError(err, mw.Window)
-			return
-		}
-		vault, err := mw.storage.CreateVault(key, name.Text)
+		vault, err := mw.storage.CreateVault(name.Text, password.Text)
 		if err != nil {
 			dialog.ShowError(err, mw.Window)
 			return
 		}
 		mw.unlockedVault[name.Text] = vault
-		mw.view.Objects[0] = newVaultView(mw, vault)
+		mw.setView(newVaultView(mw, vault))
 	})
 	btn.Importance = widget.HighImportance
 
@@ -183,18 +183,13 @@ func (mw *mainView) createVaultView() fyne.CanvasObject {
 			d.Show()
 			return
 		}
-		key, err := paw.NewKey(name.Text, password.Text)
-		if err != nil {
-			dialog.ShowError(err, mw.Window)
-			return
-		}
-		vault, err := mw.storage.CreateVault(key, name.Text)
+		vault, err := mw.storage.CreateVault(name.Text, password.Text)
 		if err != nil {
 			dialog.ShowError(err, mw.Window)
 			return
 		}
 		mw.unlockedVault[name.Text] = vault
-		mw.view.Objects[0] = newVaultView(mw, vault)
+		mw.setView(newVaultView(mw, vault))
 		mw.SetMainMenu(mw.makeMainMenu())
 	})
 	createButton.Importance = widget.HighImportance
@@ -227,7 +222,7 @@ func (mw *mainView) vaultListView() fyne.CanvasObject {
 			resource = icon.LockOutlinedIconThemed
 		}
 		btn := widget.NewButtonWithIcon(name, resource, func() {
-			mw.view.Objects[0] = mw.vaultViewByName(name)
+			mw.setView(mw.vaultViewByName(name))
 		})
 		btn.Alignment = widget.ButtonAlignLeading
 		c.Add(btn)
@@ -247,12 +242,7 @@ func (mw *mainView) unlockVaultView(name string) fyne.CanvasObject {
 	password.SetPlaceHolder("Password")
 
 	unlockBtn := widget.NewButtonWithIcon("Unlock", icon.LockOpenOutlinedIconThemed, func() {
-		key, err := paw.NewKey(name, password.Text)
-		if err != nil {
-			dialog.ShowError(err, mw.Window)
-			return
-		}
-		vault, err := mw.storage.LoadVault(key, name)
+		vault, err := mw.storage.LoadVault(name, password.Text)
 		if err != nil {
 			var invalidPasswordError *age.NoIdentityMatchError
 			if errors.As(err, &invalidPasswordError) {
@@ -262,7 +252,7 @@ func (mw *mainView) unlockVaultView(name string) fyne.CanvasObject {
 			return
 		}
 		mw.unlockedVault[name] = vault
-		mw.view.Objects[0] = newVaultView(mw, vault)
+		mw.setView(newVaultView(mw, vault))
 	})
 
 	return container.NewCenter(container.NewVBox(logo, heading, password, unlockBtn))